Add multi-site analysis scripts (CSV or Excel scenario summaries)
Significant refactor. Note that some of the utility scripts have moved to the `src` directory.

Example input and output files are included in directories of the same names for the multi-site analysis.

The multi site analysis assumes that custom load profiles are in `inputs/load_profiles`.

Added `.gitignore` and `requirements.txt`