BUG FIX: "int64 not JSON serializable"
This is a problem with the python3 version of `pandas.DataFrame.to_dict()`, which converts python data types to numpy data types (such as `numpy.int64`), which cannot be serialized by the `json` package. We use the `json` package to POST to the API, which is where the code was failing: line 45 of `post_and_poll.py`.