Update multi_site_inputs_parser.py
Lines 134-137 allows for the user to include a column in their multi-scenario input csv file titled "rate_file" that points to a specific electric rate tariff made and downloaded from the REopt Web Tool. This is in .json form and the column should contain the extension. For example, Scenario 1 would have "custom_tariff_1.json" as the element for the column titled "rate_file".