Update multi_site_inputs_parser.py
Updated the file to allow for the normal operation of the program to continue without a `urdb_json_file` field. Also added logic statements that allow for some of the scenarios to have custom electric tariff json files, but other scenarios to use URDB labels instead. Final `elif` statement breaks the program if both columns are present for any given scenario (and reports which one).